+++ lib/sanbase_web/graphql/resolvers/pricing_resolver.ex
@@ -9,9 +9,7 @@ defmodule SanbaseWeb.Graphql.Resolvers.PricingResolver do
   def subscribe(_root, %{card_token: card_token, plan_id: plan_id} = args, %{
         context: %{auth: %{current_user: current_user}}
       }) do
-    {:ok, subscription} = Subscription.subscribe(current_user.id, card_token, plan_id)
-
-    {:ok, subscription}
+    Subscription.subscribe(current_user.id, card_token, plan_id)
   end
 
   def subscriptions(%User{} = user, _args, _resolution) do
